[IMP] sale_subscription: preserve invoice day with time
If a subscription with monthly periodicity was created upon any day of
the month >= 28, it would usually get 'compressed' to the 28th of the
month since across invoicing periods it would, at some point, encounter
the month of February - e.g. a subscription start on January 31st would
have its next invoice computed to February 28th (which is normal) and
its following invoice would then be on the 28th of March (which is not
normal), then 28th of April, etc. We would prefer invoice dates to be as
close to the initial date without shrinking, e.g. Jan 31st, Feb 28th,
March 31st, April 30th, etc.

This commit introduces a mechanism to preserve the initial day of
invoicing across time. Note that if the user manually changes the date
of next invoice, then this is propagated to that mechanism (no surprise
for the end user, if they set something manually it should stick).

Add test case for this feature.