[IMP] mrp_plm: PLM Approvals improvement
When a an ECO is approved or rejected, the approval lines of the current
stage and user are updated to the last value he entered.

In order to keep a better traceability of the actions done by the users,
we'll create an approval line for each action, keep the date when it has
been approved or rejected, and log it in the chatter.

We've also added small visual improvements, to help users to know what
is the currently state of approval of the ECO.

Task-ID: 1819087

closes odoo/enterprise#2218