[IMP] project_timesheet_synchro: Animate the icon while recording time
Purpose
=======

when you use the Chrome extension, there is no way to tell whether the timer
is running or not once the extension is closed.

Specification
=============

Make the app icon in the Chrome app bar dynamic
The clock hands should be turning clockwise to show that the app is recording.
When the timer is stopped, display the current static icon.

Task ID: #1819847

Co-authored-by: Mohammed Shekha <msh@openerp.com>

closes odoo/enterprise#2295