[IMP] mrp_plm: Display ECO on products
It could be usefull to the user to know that there
exists ECO in progress on a bill of materials. It
means that the cost could change and prevent any
mistake due to it.

This commit also replace the _get_bom_reference by
a new column with the BoM version.

Task ID : 1875677

closes odoo/enterprise#2545