[FIX] sign: Avoid console error when PDF load
Currently getting console error: "Uncaught ReferenceError: jQuery is not defined" when PDF load because script loaded twice.

That's why avoid to load script twice and not removing a portion that append a 'jquery-ui.css' because main page really has no idea about what is in the iFrame - it just knows of the iFrame itself. Thus, to apply selectors css rule need to load css inside iframe.

Related to Task: 1889748

closes odoo/enterprise#2830