[IMP] mrp_workorder, quality_mrp_workorder: Change demo data for table(mto).
The demo data for table(mto) the quality points was changed.
The order of the quality points(steps) changed, but not the pdf pages they need to scroll.
So we changed the sequence of pages in pdf, and set exact worksheet page
on the steps so instruction flow in one direction.

v11 steps : https://bit.ly/2DU0qTu
v12 steps : https://urlzs.com/bTUd

task-1893502

closes odoo/enterprise#2879

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>