[IMP] sign: Allow multiple signers to sign without emails
Purpose
=======

It's already possible to send a signature request and to sign directly if you're one of the signers.

It's not possible to make a signature request to 2 people (Customer-salesman for example) and to make both people sign without clicking on the email the customer received.

Specification
=============

Add a new feature "Sign Directly" in which you choose the recipients, but intead of sending the emails,
it opens the signature request for the 1rst signer, then opens it again for the second, etc.

And finally when everyone has signed the document, send the result by email to the customer/salesman.

TaskID: 1891958