[IMP] web_studio: allow to not specify fields to export
Before this rev. when exporting a module in the `studio_customization` module,
one needed to specify both fields to export and fields not to export.

Now, only one of these two informations has to be set. The other is then
deduced with the record model fields (without magic fields).

Link to task 1891904