[IMP] web_studio: hide layout parts on text nodes
On text nodes in the Report designer, a HTML editor (summernote) is used
to design the text. Some layout parts (text decoration, colors and font style-
are thus duplicated in the components (but the one in text is more powerful).

The duplicated parts are now removed from the layout component.

The corresponding tests have been adapted and now use a <div> node.

Task 1891904