[IMP] web_studio: enable color for m2m tags
It is now possible to set the `color_field` attribute on a m2m field with
the `many2many_tags` widget.

A checkbox is displayed next to the widget selection in the sidebar.
If checked, the `color` field on the related model is selected (or `x_color`
is it doesn't exist).

Task 1888756