[IMP] sign: Manage encrypted pdf on signature
Purpose
=======

Now, if a user use a encrypted PDF, he can use it, sign it, send it,...
but when he will generate the final (signed) document to download it
(or send it by mail), he can't.

Specification
=============

When the document is signed by everyone and the document is encrypted,
just before the thankyou dialog, a new dialog ask the password to
generate the signed PDF. The new signed file is encrypted with the
same password.

A second page, managed by controller, is use is specific cases: the
last signatory bypass the password (close window for example) or for a
document already generated. We have a very simple page asking for
the password.

closes odoo/enterprise#2935