[IMP] mrp_workorder: Set a timer on the WOs tablet view based on the settings
Currently, when a user opens a WO, the time starts to be recorded
automatically. However, users don't always notice it and so they forget
to click on pause when they stop working. Thus, a timer (optional)
should be available on the WO tablet view so that the user understands
what happens.

That's why displayed a timer on the WO tablet view if timers setting is
enabled. Displayed a clock icon in orange color if user click on
'Pause'.

task-1896784

closes odoo/enterprise#2949

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>