[IMP] web_studio: always show field widget
In the View editor sidebar, the 'Widget' selection set according to the
`widget` attribute set on the field.

However, in practice, when no `widget` is set on the field, a field widget
is used (the one by default, i.e. the one with the same key as the field type).

The displayed information is not the `widget` attribute anymore but the field
widget used, because the field widget options will soon be displayed beneath it.

Task 1902867