[IMP] sign: add a selection field type in document template
Purpose
=======

Add a new "selection" field type on the documents to sign:
- The recipient is only able to choose a value among proposed options.
- The proposed options are chosen when editing the template through the configuration popover,
   any options can be added/created to any selection field.
- Options can be reused for other selection fields.

Specification
=============

A new model is added: 'sign.item.option' which represents an option of a selection field. Newly created options are global.

Options available for each selection field are also stored in a new One2Many field in 'sign.item'.

TaskID: 1893467