[ADD] voip, voip_crm: add call queue widget and mixin
Before previous voip refactoring it was possible to add a lead
to the phonecall list. This commit retore this functionnality
and add a server action to be able to add multiple records
in the next activities phonecall list. Since phonecalls are
linked to activity, it is done by creating a "Call" activity
on each record. To avoid missing anything, a UserError
will be raised if any of the record don't have a phone number.
This server action has been added on partner and leads.

The choice of making a mixin is maybe a little overkill but
will allow to easily add this behaviour on any model later.

The voip mail_activity create has been decorate by model_create_multi,
mainly to be able to create multiple record at the same time but to
send only one refresh_voip notification.

Task: 1892475