[IMP] voip: improve voip interface and manage autocall
This commit improves a little the voip pannel layout and add the
support for autocall in activities: When calling from activity
tab, each time a call is ended, the next one in the list is displayed.

Since the active call will be set has done once finished, a new option
exists in autocall mode to "postpone" the call, meaning that the activity
won't be set has done. A js counter will register witch call was already
tried in order select next one wisely and keep the already tried one
for the end.
Of course, this information will be lost if refreshing the page.

This commit also contains some cleaning and renaming, like log_note that wasn't used
and get_info() that was called multiple time, when the only information
needed was the id.

Some work should still be done to move the different behaviour to the correct place.

Task: 1892475