[IMP] account_reports, documents_account: save accounting reports.
This commit adds the possibility to save account reports as attachments in account_reports and make the bridge with the document app to save them as documents.document when this module is installed.
Some refactoring was made on accounting report buttons and localization modules to make it work flawlessly.

Was task 1889322
Superseeds PR #3077

closes odoo/enterprise#3636