[ADD] crm_helpdesk: convert leads to tickets
Purpose : leads are used as a general incoming pipeline for messages.
There should be an easy way to convert a lead into other business objects like
helpdesk tickets.

This commit adds a new module crm_helpdesk that brings a button in a lead form
view allowing quick creation of a helpdesk ticket from the lead (similar to
crm_project in community for converting leads to tasks).

Closes #3003 Task-1867574