[REF] all: update tracking parameters
This commit is linked to community PR https://github.com/odoo/odoo/pull/28430. In that PR both tracking
parameters have been merged into a single tracking parameter that is basically
the sequence enabling the tracking. Boolean value is supported as it is
considered as a default sequence of 100.

In this commit we update all enterprise addons to use the new tracking
parameter.

This merge is linked to task ID 1903814 and PR #3007.