[IMP] marketing_automation: make target model view more user friendly
Using a new tree view on ir.model we limit fields displayed in the view used
to search and find target model. Other fields are put in debug mode in order
to have them available but not in standard display mode.

Target Model field is relabeled to Target. It is easier to understand for
classic users as model is a technical word in Odoo.

This commit is linked to task ID 42548 and PR #3018.

Co-Authored-By: Kishan Gajjar <kig@odoo.com>
Co-Authored-By: Thibault Delavallée <tde@odoo.com>