[IMP] marketing_automation: allow to load sample campaigns
Purpose is to do like what is done in Sign application: a void kanban for
campaigns displays a link to load campaigns. Those are just archived records
and link is just the same action with an active flag set.

According to @est-odoo sample campaigns will come soon.

This commit is linked to task ID 42548 and PR #3018.