[REF] helpdesk: simplify default values methods
When creating a ticket, the stage and the user to assign might depends
on the team in which the ticket is created. The methods use for now are
quite fragemented and looks like spaghetti code.
This commit simplifies this business and tries to clean the code by
renaming methods with more explicit name and allowing to do it in
batch.

No functionnal changes should be observed as this is purely technical
refactoring.

Task-37561