[IMP] marketing_automation: add a link to campaign mailings
Purpose of this commit is to add a link from campaign to its mailings. As
mailings defined for marketing automation are not displayed in the classic
mass mailing application due to a domain we need a way to find them. Indeed
one could want to see its statistics or details.

We choose to keep the domain on mass mailing application in order to keep
it separated from marketing automation. A new stat button on marketing
campaign allow to jump on its mailings.

This commit is linked to task ID 42548 and PR #3018.

Co-Authored-By: Kishan Gajjar <kig@odoo.com>
Co-Authored-By: Thibault Delavallée <tde@odoo.com>