[FIX] sign: Fix undeterministic bug in salary tour
Purpose
=======

Sometime when landing on the signature page, the boxes to fill in order
to complete the signature are not displayed. This is a random bug that
crashes the salary package tour.

Specification
=============

The method waitForPDF manipulates the DOM to adds the boxes (and other things)
on the iframe.

Both Body.OnLoad() event and jQuery.ready() event will execute the javascript
code when the page is loaded.

The main differences between the two are:
- Body.Onload() event will be called only after the DOM and associated resources
  like images got loaded
- jQuery's document.ready() event will be called once the DOM is loaded i.e.,
  it wont wait for the resources like images to get loaded.

Hence, the functions in jQuery's ready event will get executed once the HTML
structure is loaded without waiting for the resources.

In our case, the method waitForPDF that is not called when the tour crashes doesn't
need the associated resources to be loaded.

So, we can replace 'load' by 'ready' and this should fix the issue as all the
executions of the tour were successful (no more undeterministic bug).

Co-authored-by: RomainLibert <rli@odoo.com>

closes odoo/enterprise#3027