[IMP] helpdesk: speed up test cases
Before this commit, all the test data were
recreated before each test.
Using the rollback is better for perfomance:
the data are created one, and then before each
test, the database state is rollbacked to it.

This commit also prepare the rewritting of a
huge part of helpdesk test cases.

Task-37561