[IMP] helpdesk: apply multi SLA per ticket
This commit makes several SLA applied to a ticket. Before, only
one SLA at a time could be applied on the ticket.

How does it work ?
On ticket creation, every existing SLA in the team configuration
is applied to the ticket, creating a list of SLA status, each with
a deadline. So a ticket, might have multiple deadline according the
SLAs applied (This status is handle by the model helpdesk.sla.status).

A SLA status is either:
- ongoing: the deadline is not passed, and the SLA stage is still not
reached
- reached: the ticket reached the SLA stage before the deadline
- failed: the deadline is passed and the ticket never reached the SLA
stage, or ticket reached the SLA stage but after the deadline.

A ticket is considered as failed when at least one of its SLA status
is failed. The ticket deadline is the deadline of the SLA currently
applied.

When the team/priority/type of a ticket is modified, all its SLA are
erased, and only the SLA of the next stages (based on sequence order)
are regenerate for the ticket.

Technical:
A special attention has been given to the performance.
- SLA status generation is done in batch
- When deleting old SLA status, and recreating new ones will only
trigger once the computation of some field on the ticket
- as we only search on ticket_id of sla.status, an index was added
- ticket creation is now using the 'create multi' decorator to be
done in batch

Limitation:
As some field are depending on "now()", we did not store them (since
the cron has been removed). Searching for failed ticket with
('sla_failed', '=', True) will lead to a new domain ['|',
('has_sla_failed', '=', True), ('sla_deadline', '>=', now())].
This might not be effective, but adding an index might solve the
problem. We will see with time.

Thanks to rco-odoo for its precious advices on performance, and
tde-banana-odoo for pointing the potential problems.

Task-37561