[IMP] helpdesk: provide SLA reporting
This commit brings a real reporting tool for
ticket failure, based on the SLA status.
This allow the user to find the SLA failed, the
problematic stage, the 'not so good' employees,
the ratio of failing ticket, ... thanks to a
new SQL view.

This commit also bring 2 predefined filters
as data to help user.

Task-37561