[ADD] approvals: New module to manage approvals mechanism
Purpose
=======

Develop a new application to manage approvals workflow: business trips, out of office,
overtime, borrow items, general approvals, procurements, contract approval, etc.

There is already a pseudo-approval-mechanism for some models such as expenses, leaves, ...
When a new record is created, a next activity is generated for the manager to make him
aware of the record creation.

The goal of this new application is to provide a generic approvals mechanism for non
existing Odoo models.

As an example, at Odoo SA, we could use the approval module for employees to:

- Ask purchase requests
- Ask advance payment for a roadshow
- Validate a marketing campaign budget

Specification
=============

An approval category can specify the mandatory fields to fill on a request, like
an attachment, dates (from/to), an amount, ...

Some data are already created for the most common approvals category.

When confirming an approval request, some next activities are created for
all the approvers. They can accept or refuse it. The next activity is logged on
the chatter and the employee is warned by email/notification about the decision
of each approver.

closes odoo/enterprise#3249