[IMP] sign : add next activity on sign request
When you ask for a signature, To Do "Sign document" activity is
created for each internal users that have to sign the document
and that are at least sign officer.
Signing the document marks the activity as done, but marking
the activity as done has no impact on the document.

Task ID : 1912199
Closes PR #3161