[REF] web(site)_studio: adapt code to web refactoring
A community commit (see odoo/odoo#29106) moved the ControlPanel
from the ActionManager to the controllers. A lot of logic has been
moved from the ActionManager to where it belongs. This is for
example the case for the 'env' (that was previously set on the
action). This had an large impact on Studio, which needed to be
adapted.

Part of task 1893568

Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>