[IMP] documents*: improves the demo data & action UI
This commit adds and changes several demo of documents
to provide a better demonstration workflow.

In addition, it changes how the Actions are displayed:
before this commit, all of them were displayed in
a list view accessible via a menuitem (now in debug mode);
after this commit, a stat button is available on the Folder's
form view to display all the actions of the selected folder.

task: 1904778
closes: odoo/enterprise#3247