[IMP] documents: adds a global view
After this commit, the default kanban attachment view is the
document's kanban view.

When the view is opened, no folder is selected by default:
all the attachments that are inside all available folders are displayed;
it is now also possible to click a "all" tab to go back to
that global view.

Task: #1905664

closes odoo/enterprise#2995