[FIX] website_sale_coupon: no crash allowed in tour
The tour shop_sale_coupon was changed to use the testUtils.dom.click
method when we introduced the new test utility functions.  Probably
because the filename ends with 'test.js'.  Obviously, this is not
correct, because the testUtils function is not imported.

To prevent issues in the future, we simply remove the two problematic
calls.  Another solution was to import the testUtils module, however, I
think that it is simpler not to use those functions at all in tours,
because tours can be executed in the web client/website, and in that
case, they do not have access to the test utility functions.

closes odoo/enterprise#3239