[FIX] documents_account: prevents duplicate invoices
Before this commit, when creating an invoice with a workflow action
on a document that was already attached to an invoice resulted in a duplicate.

After this commit, when the action is applied, if the document is already
attached to an invoice, it only opens the appropriate view to this invoice.

Task: #1916433

closes odoo/enterprise#3274