[REF] hr_appraisal: adapt code to survey invite update of templates
Templates used in invite wizard should now be based on the answer (user_input)
model. It has be done in order to ease the rendering of invite email body and
more specifically access link with token depending on target answer. In this
commit we update code, data and demo accordingly.

In this commit we also rename the mail_template_id field into
survey_template_id in order to be more specific. It is not a generic template
used in appraisal but a template effectively linked to the invite process.

This commit is linked to task ID 1911586 and PR #3197.