[REF] hr_appraisal: correctly store and handle appraisal on survey
Purpose of this commit is to avoid relying on context to store and propagate
the appraisal id on invite and user answers. We now add a real appraisal_id
field on invite wizard that is propagated to created answers accordingly when
sending the invites.

This commit is linked to task ID 1911586 and PR #3197.