[IMP] marketing_automation: propagate domain to children activities
The domain that is used on an activity isn't its domain but the product of all
its parent in the activity hierarchy, starting from the marketing campaign's
domain.
This commit renames the `domain` field to `activity_domain`, keeping the
functionality as before.
The field domain is used as a compute field that holds the product of all
inherited domains and the activity_domain.

opw 1889719

closes odoo/enterprise#3308