[IMP] account_sepa: EPC QR code on supplier payments
In Europe, there is the EPC QR Code that can be used to record a payment with mobile apps. In the wizard to record a payment to a supplier, we should have the ability to show the EPC QR Code. That way, the user can scan it on the screen with his mobile and pay right away.

Was task ID #1918423

closes odoo/enterprise#3944

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>