[IMP] hr_appraisal: add appraisal category for surveys
Purpose of this commit is to add a new survey category for appraisals. That
way surveys created and/or used through appraisal process are correctly
set in their own category allowing to filter them. When creating a new
appraisal only surveys linked to appraisal process are shown. Moreover
default values are propagated when creating new surveys so that they are
correctly configured by default as appraisal and invite-only surveys.

This commit is linked to task ID 1911586 and PR #3197.