[IMP] web_mobile, ocn_client: notify recipients only if needed
(1)
Before this commit, ocn_client was automatically installed at database
creation even if push notifications was disabled.
As a consequence, useless Python code was executed in ocn_client when
mail message was created (_notify_compute_recipients).

Now, ocn_client will be installed only if "Push Notifications" setting
is enabled. For that we had to move this setting in web_mobile.

(2)
Before this commit, requests was done to fetch partners to send
notification even if no FCM project ID was configured.
This project was set when "Push Notifications" was checked.

To avoid these unnecessary queries, we set FCM project ID only if
at least one device has been logged to the mobile app and we check
that this project is properly set before fetching partners.
We consider that it's not necessary to send notifications if there
is no mobile device registered.

Task ID : 1966283