[FIX] l10n_mx_edi: Payments for invoices with multicurrency is supported
The PAC makes some validations in the CFDI documents, to ensure that the
document information is correct and that the CFDI could be send to the
SAT to sign.

One of this validations is the next:
The sum of the values in ``ImpPagado`` of each invoice must be less than
or equal that the attribute ``Monto``.

And when the payment is different that the invoice that is paid, is
converted the invoice amount in the payment currency, based on the
attribute ``TipoCambioDR``, this, to make the check correctly.

Then, in the next case:
Rates:
MXN = 1
USD = 20

Create 2 invoices:
Invoice 1 for 100 MXN
Invoice 2 for 100 USD

Now make a payment from the bank statement for 105 USD

The CFDI for the payment returns the next error:
Message: La suma de los valores registrados en el campo ImpPagado de los
apartados DoctoRelacionado no es menor o igual que el valor del campo
Monto.

Why?
The ``TipoCambioDR="1.904762"`` in the CFDI, then:
100 USD + (100 MXN / 1.904762 = 52.4999973) = 152.4999973, and is
different to 105

Now, with this fix, the CFDI is the next TipoCambioDR="20.000000":
100 USD + (100 MXN / 20 = 5) = 105

Test added for this new user case

closes odoo/enterprise#5600

X-original-commit: 70bd13e9cdb35d01ad95b0937f730fe2c7b52cb8
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>