[IMP] timesheet_grid: Change behaviour of onchange parent_id
Purpose
=======

Currently, when an employee is assigned to a new manager,
his timesheet responsible is also set to this new manager.
However, assigning a new manager to an employee does not
always mean that the timesheet responsible should also change.

Specification
=============

When changing the employee's manager, the timesheet responsible
should be changed only if the field was not set or if the
responsible is the previous manager.
In that case, the timesheet responsible should be set to the
new manager.
Otherwise (timesheet responsible is an other person than the
previous manager), leave the responsible unchanged.