[IMP] helpdesk: remove cron for computed field depending on now()
Helpdesk was built with some computed fields depending on now() (current
time). For some reason those field were stored (probably to ease the
search). To fix the "update problem", a cron was added to recompute
those fields.
Concerned field are 'assign_hours' and 'close_hours'.

This can lead to some issues in term of performance. This cron means:
- an inactive database will still recompute those fields and use CPU, ...
- the more ticket are open, the more the cron will take time and slow
down the system
- when the cron is running, it put a "shared lock" on the record to
recompute (all open ticket). During that time, if a user is working on
a ticket, he will put another "shared lock" on it. Those "shared lock"
will be removed by an "explicit lock" when the record will be modify
(UPDATE and COMMIT, by cron or user). In that case, the other "locker"
will have its transaction rollbacked, and retry by odoo. It is obvious
that this can lead to serialization failure will happen and that some
operation can not be completed, as they might be in concurrency.

This commit removes this cron to avoid those problems. We also realize
the compute method were wrong to compute those fields. So, the commit
also fixes the implementation of the compute methods:
- assign_hours is relevant only when the ticket was assigned. Otherwise
its value be null. It now takes the working calendar into account.
- close_hours is relevant only when the ticket is closed. Otherwise its
value is null. It now takes the working calendar into account.
- open_hours: new field created (and depending on NOW()), in order to
have the lifetime duration as 'close_hours' was wrongly trying to do.
This field does not take the working calendar into account, as it is
now-depending and we might want to search on it.

Task-37561