[IMP] hr_appraisal, hr_documents, timesheet_grid: integration to employee profile
This commit is part of the new employee profile (Task 1913089)

hr_appraisal
============

This commit integrate appraisals to the new employee profile.
It adds a button to allow a user to request an appraisal (see below)
Also adds a stat button displaying the date of the next appraisal.
The button redirects to the employee's appraisals.

An employee should be allowed to request an appraisal himself
through his profile.

Additionnal settings are added to control requests.
The maximum and minimum appraisal frequency.
e.g.:
  max = 6 month and
  min = 18 month
  -> An employee cannot request a new appraisal if 6 month
     haven't passed since the last one.
  -> A new appraisal will be created if 18 month (minus 7 days)
     have passed since the last appraisal.

hr_documents
============

New bridge module allowing to integrate Documents to the new employee profile.
Add a stat button on the employee profile linking to its documents.

timesheet_grid
==============

Move timesheet responsible in the Managers section of the employee profile