[IMP] documents: adds folder groups and improves share access check
This commit changes how the share access is checked and how their
documents are accessed: it is now done through the methods on `documents.share`
and evaluated based on the current access rights of the create_uid of
the `documents.share`.

This commit also allows documents managers to add readonly groups to
folders and to restrict the content by document `owner_id`.

adds two new fields (`read_group_ids` and `user_specific`)
to `documents.folder` and one new computed field (`can_upload`) to `documents.share`

Task: #1915003

closes odoo/enterprise#3411

Signed-off-by: Martin Geubelle (mge) <mge@openerp.com>