[FIX] documents: corrects filter fields
Following the refactor (20f5e74a8620ae6629327ef833bedfea0728c923), the field `name` is used for both
binary attachments and URL/requests, so the `datas_fname` filter
and the custom filter for `name` are not necessary anymore.

This commit replaces the custom `name` filter and the `datas_fname` filter
by a single `name` filter.

closes odoo/enterprise#3382