[IMP] *account_taxcloud: hide taxes when using Taxcloud
The generated tax names contain an estimate of the percentage. This
leads to issues. E.g.

|---------+------------+-------------|
| Product | Unit price | Taxes       |
|---------+------------+-------------|
| A       |      10.00 | Tax 7.700 % |
| A       |      55.00 | Tax 7.745 % |
| A       |     100.00 | Tax 7.750 % |
|---------+------------+-------------|

These 3 tax rates are really the same, but to make the tax amount
correct after rounding Odoo calculates 3 different taxes. This is
confusing to customers.

Instead always show the tax amount per line and hide the taxes if the
fiscal position uses Taxcloud. This is in line with how the Taxcloud
API is meant to be used. They just return only the amount and if the
user needs details they should check on their Taxcloud account.

Note that the issue does not appear on reports because these generated
taxes are grouped under a generic name.

Related to odoo/odoo#18084