[IMP] marketing_automation: correctly set traces on error when failed emails
Currently if there is no crash when sending emails all traces are considered
as sent and successful. However in some cases emails are not sent without
having an exception. For example mass mailing set duplicate emails to send
as ignored and traces should be updated accordingly.

In this commit we fetch statistics of emails that are in exception or ignored
after sending the emails of mass mailing activities. Traces linked to those
statistics are updated accordingly.

It allow to have more meaningful results and statistics in marketing
automation activities and have more information about what happened.

This commit is linked to task ID 1924711 and PR #3391.