[IMP] marketing_automation: improve mail statistics form view
Purpose is to ease daily use of mass mailing and marketing automation by
adding relevant fields on statistics form view. As this is a technical model
when user go on statistics form view it is better to display all available
information.

In community several fields are added. In this commit we simply add the
trace linked to the statistics.

This commit is linked to task ID 1924711 and PR #3391.