[IMP] marketing_automation: add bounced statistics on email activities
Currently statistics for sent, replied and opened are displayed on activities
kanban cards. Bounce stats is missing as this information is usefull but not
displayed. Indeed sent is the number of sent emails that can after be opened,
replied or bounced. All those statistics should be displayed.

This commit is linked to task ID 1924711 and PR #3391.