[IMP] marketing_automation: add click statistics on marketing campaigns
Purpose of this commit is to add a stat button on marketing campaign that
redirects on tracker statistics. It allows to see at a glance which links
were clicked and see more in details the links / mailings that are more
efficient.

This commit is linked to task ID 1924711 and PR #3391.