[IMP] sale_timesheet_enterprise: add timer
Add a new feature for projects with timesheets: a START button to launch
timer on a task while working on it. The working time is displayed on the
form and, when STOP button is clicked, added as a timesheet in the task
with the possibility to edit the description.

Task-1928442 Closes #3639