[IMP] web_mobile: control panel appears on scroll up
When user go down in the page control panel disappears and
when user go up in the page control panel progressively appears on top.

When the scroll position of the page is after the control panel
we add o_mobile_sticky in class list of control panel.

On every scroll event we recompute the top attribute of the control panel
to partially show or hide depending if you are scroll up or down.
By doing this we prevent the control panel from suddenly disappearing.

closes odoo/enterprise#3538