[FIX] documents: fixes the file upload
This commit changes how files are uploaded from the
front-end.

Reason: the previous implementation was limiting the upload
and was causing crashes for large file uploads.

This commit also sets the default order of documents
as descending by id (most recent first), to keep the same behaviour
as the 12.0 version based on 'ir.attachment`.

closes odoo/enterprise#3545

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>