[FIX] voip: inform user when call activity type is wrongly configured
`create_call_in_queue` highly relies on the fact that activity type
`mail_activity_data_call` has 'phonecall' as category. It isn't always the case
since activity type can be edited by users.

In this case, it is possible that a user receives a 'record may have no phonecall'
error message when it is actually linked to activity_type configuration.

This commit adds a better error message in this case,
(and hopefully prevents support team to waste time on related tickets)

closes odoo/enterprise#3546