[IMP] account_predictive_bills: make tests work with l10n_be
In account predictive bills we try to match the invoice line to a
journal using the full text search engine of postgresql.
When testing the feature, we use test journals to check if they are
indeed matched, this works until you try to install l10n_be in order to
test it.
In this case, you introduce an journal with exactly the same name than
the test journal which will obviously be matched and as it will have
been created before the test journal it will be taken first (order by
id).
This will lead to the test crashing because it doesn't match the
expected journal.

By changing those names in the test, we can avoid this conflict situation.

Task 1913733

closes odoo/enterprise#3486