[IMP] web_studio: lazy load studio and wysiwyg assets
Issue: wysiwyg asset slow down the loading of the website, error
inadvertently introduced: https://github.com/odoo/odoo/pull/29775

The fix of this issue in community makes it necessary to put parts of
the editor in asynchroneous load.
Studio has a lot of includes that cannot be lazy loaded, thus we need
the 2 asset bundles.

closes odoo/enterprise#3552