[IMP] sign,*: toggle sign button depending on signature state
* = hr_contract_salary

Before this commit, it was possible for the user to click on the Sign button
even when the signature area was empty, which would lead to an error message.
Now we disable the button when the area is empty.

With this change, we can adapt the hr_contract_salary tour to make it more
deterministic by correctly clicking on the Sign button only after the signature
area has been filled.

Thanks to @d-fence for the idea in PR #3281.

Related to community PR: https://github.com/odoo/odoo/pull/30653

PR: #3535